+++ src/Discretise/boundary_conditions/fixedTemperature.jl
@@ -1,5 +1,6 @@
 export FixedTemperature
 export Enthalpy
+export IE
 
 """
     FixedTemperature(name::Symbol, model::Enthalpy; T::Number)
@@ -28,14 +29,36 @@ Adapt.@adapt_structure FixedTemperature
 end
 Adapt.@adapt_structure FixedTemperatureValue
 
-# Temperature models (Enthalpy only for now!)
+# Temperature models
 @kwdef struct Enthalpy{C,F}
     cp::C
     Tref::F
 end
 
-# API-Level constructor for FixedTemperature
+"""
+    IE{C,F}
+
+BC converter for internal energy formulation: `e = cv*(T - Tref)` where `cv` is
+the specific heat at constant volume.
+
+# Fields
+- `cv`: Specific heat at constant volume.
+- `Tref`: Reference temperature.
+
+# Example
+    FixedTemperature(:inlet, T=300.0, IE(cv=718.0, Tref=288.15))
+"""
+@kwdef struct IE{C,F}
+    cv::C
+    Tref::F
+end
+
+# API-Level constructors for FixedTemperature
 FixedTemperature(name, model::Enthalpy; T) = begin
+    FixedTemperature(name, FixedTemperatureValue(T=T, energy_model=model))
+end
+
+FixedTemperature(name, model::IE; T) = begin
     FixedTemperature(name, FixedTemperatureValue(T=T, energy_model=model))
 end
 
@@ -45,6 +68,14 @@ end
     Tref = model.Tref
     h = cp*(T - Tref)
     return h
+end
+
+# Conversion temperature to internal energy
+@inline (model::IE)(T) = begin
+    cv = model.cv
+    Tref = model.Tref
+    e = cv*(T - Tref)
+    return e
 end
 
 adapt_value(value::FixedTemperatureValue, mesh) = begin
@@ -82,10 +113,16 @@ end
 
 @define_boundary FixedTemperature Divergence{BoundedUpwind} begin
     (; T, energy_model) = bc.value
-    flux = term.flux[fID]
-    ap = term.sign*(flux)
+    # flux = term.flux[fID]
+    # ap = term.sign*(flux)
+    # ac = max(-ap, 0.0)
+    # 0.0, -ap*h
+    
+    ap = term.sign*(term.flux[fID])
+    ac = max(-ap, 0.0)
+    an = -max(-ap, 0.0)
     h = energy_model(T)
-    -flux, -ap*h
+    ac, -an*h
 end
 
 @define_boundary FixedTemperature Laplacian{Linear} begin